moved log into summary tabs
