+++ web/package.json
@@ -70,10 +70,11 @@
     "@types/node": "^12.11.1",
     "codelyzer": "^5.1.2",
     "conventional-changelog-cli": "1.3.4",
+    "electron": "^9.0.3",
     "husky": "0.13.3",
     "jasmine-core": "2.6.4",
     "jasmine-spec-reporter": "4.1.1",
-    "karma": "1.7.1",
+    "karma": "^1.7.1",
     "karma-chrome-launcher": "2.1.1",
     "karma-cli": "1.0.1",
     "karma-coverage-istanbul-reporter": "1.3.0",
