+++ web/src/app/pages/detail/detail.component.html
@@ -13,4 +13,7 @@
       [(ngModel)]="model"
     ></ng-monaco-editor>
   </nb-tab>
+  <nb-tab tabTitle="Logs">
+    <ngx-logs></ngx-logs>
+  </nb-tab>
 </nb-tabset>
