+++ web/src/app/pages/detail/logs/log.ts
