+++ web/src/app/pages/detail/logs/logs.component.html
@@ -13,6 +13,6 @@
 <br>
 <div style="background: white;">
   <ul *ngFor="let log of logs">
-    <p *ngIf="isActive(log.level)" style="color: black; white-space: pre-wrap;"> {{log.log}}</p>
+    <p *ngIf="isActive(log.level)" style="color: black; white-space: pre-wrap;"> {{ log.log }}</p>
   </ul>
 </div>
