+++ web/src/app/pages/detail/logs/logs.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
-import { InstanceService } from '../../instances/instance.service';
+import { InstanceService } from '../../../instances/instance.service';
 import 'style-loader!angular2-toaster/toaster.css';
 import {
   NbComponentStatus,
